op-deployer,ctb: Migrate DeployProxy to new script API
