+++ packages/contracts-bedrock/test/opcm/DeployProxy.t.sol
@@ -3,23 +3,23 @@ pragma solidity 0.8.15;
 
 import { Test } from "forge-std/Test.sol";
 
-import { DeployProxy2 } from "scripts/deploy/DeployProxy2.s.sol";
+import { DeployProxy } from "scripts/deploy/DeployProxy.s.sol";
 
-contract DeployProxy2_Test is Test {
-    DeployProxy2 deployProxy;
+contract DeployProxy_Test is Test {
+    DeployProxy deployProxy;
 
     // Define default input variables for testing.
     address defaultProxyAdmin = makeAddr("ProxyAdmin");
 
     function setUp() public {
-        deployProxy = new DeployProxy2();
+        deployProxy = new DeployProxy();
     }
 
-    function testFuzz_run_memory_succeeds(DeployProxy2.Input memory _input) public {
+    function testFuzz_run_memory_succeeds(DeployProxy.Input memory _input) public {
         vm.assume(_input.owner != address(0));
 
         // Run the deployment script.
-        DeployProxy2.Output memory output = deployProxy.run(_input);
+        DeployProxy.Output memory output = deployProxy.run(_input);
 
         // Assert inputs were properly passed through to the contract initializers.
         vm.prank(_input.owner);
@@ -27,7 +27,7 @@ contract DeployProxy2_Test is Test {
     }
 
     function test_run_nullInput_reverts() public {
-        DeployProxy2.Input memory input;
+        DeployProxy.Input memory input;
 
         input = defaultInput();
         input.owner = address(0);
@@ -35,7 +35,7 @@ contract DeployProxy2_Test is Test {
         deployProxy.run(input);
     }
 
-    function defaultInput() internal view returns (DeployProxy2.Input memory input_) {
-        input_ = DeployProxy2.Input({ owner: defaultProxyAdmin });
+    function defaultInput() internal view returns (DeployProxy.Input memory input_) {
+        input_ = DeployProxy.Input({ owner: defaultProxyAdmin });
     }
 }
